fix(dagster): move some things around
