+++ warehouse/oso_dagster/definitions/common.py
@@ -1,5 +1,9 @@
+import typing as t
+
 from dagster import Definitions
 from oso_dagster.factories.common import AssetFactoryResponse
+
+from .resources import default_resource_registry
 
 
 def load_definitions_with_asset_factories(
@@ -12,3 +16,14 @@ def load_definitions_with_asset_factories(
         sensors=asset_factories.sensors,
         **kwargs,
     )
+
+
+def run_with_default_resources[T](
+    func: t.Callable[..., T],
+    **kwargs: t.Any,
+) -> T:
+    """Run a function with the default resources."""
+    registry = default_resource_registry()
+    resources = registry.context()
+
+    return resources.run(func, **kwargs)
